Exception handling

+++ src/main/java/com/devsuperior/clientsp/services/ClientService.java
@@ -3,11 +3,15 @@ package com.devsuperior.clientsp.services;
 import com.devsuperior.clientsp.dto.ClientDTO;
 import com.devsuperior.clientsp.entities.Client;
 import com.devsuperior.clientsp.repositories.ClientRepository;
+import com.devsuperior.clientsp.services.exceptions.DatabaseException;
+import com.devsuperior.clientsp.services.exceptions.ResourceNotFoundException;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -21,7 +25,8 @@ public class ClientService {
 
     @Transactional(readOnly = true)
     public ClientDTO findById(Long id){
-        Client client = repository.findById(id).get();
+        Client client = repository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Recurso não encontrado."));
         return modelMapper.map(client, ClientDTO.class);
     }
 
@@ -47,9 +52,16 @@ public class ClientService {
         return modelMapper.map(client, ClientDTO.class);
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-        repository.deleteById(id);
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso não encontrado.");
+        }
+        try {
+            repository.deleteById(id);
+        }catch(DataIntegrityViolationException e){
+            throw new DatabaseException("Falha de integridade referencial.");
+        }
     }
 
     public void convertClientToDTO(Client client, ClientDTO dto){
